feat: adds new action and requirement system
Implements a new action and requirement system for menu items.

This change introduces a more flexible way to define actions
that can be performed when a menu item is clicked, as well as
requirements that must be met for a menu item to be visible or
clickable.

Adds new actions such as:
- Sending a message to a player
- Closing the inventory

Adds support for "deny-actions" in requirements, allowing to
specify actions that will be performed if the requirement is not
met.